% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaCompData}
\alias{alpinaCompData}
\title{Arabis alpina floral scent compounds}
\format{
A data frame with 15 rows and 3 columns. Each row is a compound.
First column is a common name of the compound, second column is the
SMILES (Simplified Molecular-Input Line-Entry System) specification,
third column is the InChIKey (International Chemical Identifier).
}
\source{
Petren H, Torang P, Agren J, Friberg M. 2021.
Evolution of floral scent in relation to self-incompatibility and capacity
for autonomous self-pollination in the perennial herb \emph{Arabis alpina}.
Annals of Botany 127: 737-747.
}
\usage{
alpinaCompData
}
\description{
A dataset listing the compounds in \code{\link{alpinaSampData}}.
}
\keyword{datasets}
