% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaCompDis}
\alias{alpinaCompDis}
\title{Arabis alpina floral scent compound dissimilarity matrix}
\format{
A 15x15 compound dissimilarity matrix.
}
\usage{
alpinaCompDis
}
\description{
A matrix with compound dissimilarities calculated using
\code{\link{compDis}} with \code{type = "PubChemFingerprint"},
for the compounds in \code{\link{alpinaCompData}}.
}
\keyword{datasets}
