% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaCompSubset}
\alias{alpinaCompSubset}
\title{Arabis alpina floral scent compounds}
\format{
A data frame with 15 rows and 3 columns. Each row is a compound.
First column is a common name of the compound, second column is the
SMILES (Simplified Molecular-Input Line-Entry System) specification,
third column is the InChIKey (International Chemical Identifier).
}
\source{
\url{https://doi.org/10.1093/aob/mcab007}
}
\usage{
alpinaCompSubset
}
\description{
A dataset listing the compounds found in \code{\link{alpinaDataSubset}}.
}
\keyword{datasets}
