% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaDataSubset}
\alias{alpinaDataSubset}
\title{Arabis alpina floral scent data}
\format{
A data frame with 87 rows and 16 columns. Each row is a sample,
first column is population, the other columns are scent compounds.
}
\source{
\url{https://doi.org/10.1093/aob/mcab007}
}
\usage{
alpinaDataSubset
}
\description{
A dataset with (proportional) floral scent data from three populations of
the plant species Arabis alpina.
}
\keyword{datasets}
