% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaMolNet}
\alias{alpinaMolNet}
\title{Arabis alpina floral scent molecular network}
\format{
A \code{tbl_graph} object with 15 nodes and 56 edges.
}
\usage{
alpinaMolNet
}
\description{
A molecular network. Generated by the \code{\link{molNet}} function
used on the \code{\link{alpinaCompDis}} dataset, with
\code{cutOff = 0.75}.
}
\keyword{datasets}
