% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaNPCTable}
\alias{alpinaNPCTable}
\title{Arabis alpina floral scent NPClassifier table}
\format{
A dataframe with 15 compounds and their NPClassifier classifications.
}
\usage{
alpinaNPCTable
}
\description{
A table with the NPClassifier pathways, superclasses and classes,
along with the compound names, smiles and inchikey. Generated by
the \code{\link{NPCTable}} function used on the
\code{\link{alpinaCompData}} dataset.
}
\keyword{datasets}
