% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaPopData}
\alias{alpinaPopData}
\title{Arabis alpina populations}
\format{
A data frame with 87 rows and 1 column. Each row represents
the population each sample in \code{\link{alpinaSampData}} comes from
(It8, S1 or G1).
}
\source{
Petren H, Torang P, Agren J, Friberg M. 2021.
Evolution of floral scent in relation to self-incompatibility and capacity
for autonomous self-pollination in the perennial herb \emph{Arabis alpina}.
Annals of Botany 127: 737-747.
}
\usage{
alpinaPopData
}
\description{
A dataset listing what population each sample
in \code{\link{alpinaSampData}} comes from.
}
\keyword{datasets}
