% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaSampData}
\alias{alpinaSampData}
\title{Arabis alpina floral scent data}
\format{
A data frame with 87 rows and 16 columns. Each row is a sample,
first column is population, the other columns are scent compounds.
}
\source{
Petren H, Torang P, Agren J, Friberg M. 2021.
Evolution of floral scent in relation to self-incompatibility and capacity
for autonomous self-pollination in the perennial herb \emph{Arabis alpina}.
Annals of Botany 127: 737-747.
\url{https://doi.org/10.1093/aob/mcab007}
}
\usage{
alpinaSampData
}
\description{
A dataset with (proportional) floral scent data from three populations of
the plant species Arabis alpina.
}
\keyword{datasets}
