% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaSampDis}
\alias{alpinaSampDis}
\title{Arabis alpina floral scent sample dissimilarity matrix}
\format{
A 87x87 sample dissimilarity matrix.
}
\usage{
alpinaSampDis
}
\description{
A matrix with sample dissimilarities calculated from
\code{\link{alpinaSampData}} and \code{\link{alpinaCompDis}}
using \code{\link{sampDis}} with \code{type = "GenUniFrac"} and
\code{alpha = 0.5}.
}
\keyword{datasets}
