% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiv.R
\name{calcDiv}
\alias{calcDiv}
\title{Calculate various diversity/evenness measures}
\usage{
calcDiv(sampleData, compDisMat = NULL, type = "HillDiv", q = 1)
}
\arguments{
\item{sampleData}{Dataframe with samples as rows and compounds as columns.}

\item{compDisMat}{Compound dissimilarity matrix, as calculated by
\code{\link{compDis}}. Has to be supplied for
calculations of Functional Hill diversity, Rao's Q and FAD.}

\item{type}{Type(s) of diversity or evenness to calculate. Any of
\code{"HillDiv", "FuncHillDiv", "Shannon", "Simpson", "PielouEven",
"HillEven", "RaoQ", "FAD"}.}

\item{q}{Diversity order to use for (Functional) Hill diversity. q should
be equal to or larger than zero. This parameter determines the sensitivity
of the (Functional) Hill diversity measure to the relative frequencies
of compounds. For q = 0 compound proportions are not taken into account.
For q = 1 compounds are weighed according to their proportion in the sample.
For q = 2, more weight is put on compounds with high proportions.}
}
\value{
Data frame with calculated diversity/evenness for each sample.
}
\description{
Function to calculate Hill diveristy, Functional Hill diversity,
Shannon's diveristy, Simpson diveristy, Pielou's evenness, Hill evenness,
Rao's Q and Functional Attribute Diversity (FAD).
}
\details{
Add details here
Hill diversity:
Functional Hill diversity: With compound matrix.
Shannon's diversity:
Simpson diversity:
Pielou's evenness:
Hill evenness:
Rao's Q: With compound matrix.
FAD: Functional Attribute Diversity.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
calcDiv(sampleData = minimalSampData)
calcDiv(sampleData = minimalSampData, type = c("HillDiv", "HillEven"))
calcDiv(sampleData = minimalSampData, compDisMat = minimalCompDis,
type = "FuncHillDiv", q = 2)
}
\references{
Chao's and other papers
Chao, A., C.-H. Chiu, and L. Jost. 2014. Unifying Species Diversity,
Phylogenetic Diversity, Functional Diversity, and Related Similarity
and Differentiation Measures Through Hill Numbers. Annual Review of
Ecology, Evolution, and Systematics 45:297–324.
}
