% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDivProf.R
\name{calcDivProf}
\alias{calcDivProf}
\title{Calculate diversity profile}
\usage{
calcDivProf(sampleData, compDisMat = NULL, qMin = 0, qMax = 3, step = 0.1)
}
\arguments{
\item{sampleData}{Data frame with samples as rows and compounds as columns.}

\item{compDisMat}{Compound distance matrix, as calculated by
\code{\link{compDis}}. Has to be supplied for calculations of
Functional Hill diversity.}

\item{qMin}{Minimum value of q}

\item{qMax}{Maximum value of q}

\item{step}{Step by which q will be calculated between qMin and qMax}
}
\value{
List with a diversity profile data frame with samples as rows
and the (Functional) Hill diversity at different q values as columns;
and values for qMin, qMax and step.
}
\description{
Function to calculate a "diversity profile", i.e. calculate (Functional)
Hill diversity for multiple values of q.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
calcDivProf(sampleData = minimalSampData)
calcDivProf(sampleData = minimalSampData, compDisMat = minimalCompDis,
qMin = 1, qMax = 2, step = 0.2)
}
