% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateQ}
\alias{calculateQ}
\title{Rao's Q}
\usage{
calculateQ(data, Dij)
}
\arguments{
\item{data}{Dataset. Single row in dataframe. Proportional or absolute data.}

\item{Dij}{Dissimilarity matrix.}
}
\value{
Value of Rao's Q.
}
\description{
Function to calculate Rao's Q.
Note that this is done for one sample (row).
}
