% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdiv.R
\docType{package}
\name{chemdiv}
\alias{chemdiv}
\title{chemdiv: A package for analysing phytochemical diversity}
\description{
chemdiv is an R package for analysing phytochemical data. The package
includes a number of functions that enables quantification and
visualization of phytochemical diversity and dissimilarity for any type of
phytochemical (and similar) samples, such as herbivore defence
compounds, volatiles or similar. Importantly, calculations
of diversity and dissimilarity can incorporate biosynthetic
and/or structural properties of the phytochemical compounds, resulting
in a more complete quantification of diversity and dissimilarity.
Functions in the R-package will work best for sets of data, commonly
generated by chemical ecologists using GC-MS, LC-MS or similar, where all
or most compounds in the samples have been confidently identified.
}
\details{
Two datasets are needed to use the full set of analyses
included in the package. The first dataset should contain data on the
relative proportions of different compounds (columns) in
different samples (rows). Note that all calculations of diversity, and
most calculations of dissimilarity, can only be performed on relative,
rather than absolute, concentrations.
The second dataset should contain, in each of three columns in a data frame,
the compound name, SMILES and InChIKey IDs of all the compounds
present in the first dataset. SMILES and InChIKey are chemical identifiers
that are easily obtained for each compound by searching for it in
PubChem \url{https://pubchem.ncbi.nlm.nih.gov/}, or by using various
automated tools such as the PubChem Identifier Exchange Service
\url{https://pubchem.ncbi.nlm.nih.gov/idexchange/idexchange.cgi} or
The Chemical Translation Service \url{https://cts.fiehnlab.ucdavis.edu/}.
The user is intentionally required to do this manually to ensure correctness,
as lists of compounds very often contain compounds wrongly named,
wrongly formatted, under various synonyms etc. which prevents easy automatic
translation of compound names to SMILES and InChIKey.

Once the two datasets are prepared, these should be imported into R
as separate data frames, and all analyses in the package can then be
performed, in largely the same order as they appear in the list below.
}
\section{Data format checks}{

\code{\link{chemDivCheck}}
}

\section{Compound classification and dissimilarity}{

\code{\link{NPCTable}}
\code{\link{compDis}}
}

\section{Diversity calculations}{

\code{\link{calcDiv}}
\code{\link{calcBetaDiv}}
\code{\link{calcDivProf}}
}

\section{Sample dissimilarities}{

\code{\link{sampleDis}}
}

\section{Molecular network and properties}{

\code{\link{molNet}}
}

\section{Chemodiversity and network plots}{

\code{\link{molNetPlot}}
\code{\link{chemDivPlot}}
}

\references{
A bunch probably
}
\author{
List authors here
}
