% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemDivCheck.R
\name{chemDivCheck}
\alias{chemDivCheck}
\title{Check data formatting}
\usage{
chemDivCheck(compoundData, sampleData)
}
\arguments{
\item{compoundData}{Data frame with the chemical compounds of interest.
Should have a column named "compound" with common names, a column named
"smiles" with SMILES IDs for the compounds, and a column named "inchikey"
with the InChIKey IDs for the compounds.}

\item{sampleData}{Data frame with the relative concentration of each
compound (column) in every sample (row).}
}
\value{
One or several messages pointing out problems with data formatting,
or a message informing that the datasets are correctly formatted.
}
\description{
This function checks so that the datasets used by functions in
the package are correctly formatted.
}
\details{
The function performs a number of checks on the two main datasets used
as input data, to make sure datasets are formatted in a way suitable
for the other functions in the package. This should make it easier for
users to correctly format datasets before starting with analyses.

Two datasets are needed to use the full set of analyses included in
the package, and can be checked for formatting issues.
The first dataset should contain data on the proportions
of different compounds (columns) in different samples (rows).
Note that all calculations of diversity, and most calculations of
dissimilarity, can only be performed on relative, rather than absolute,
concentrations. The second dataset should contain, in each of three
columns in a data frame, the compound name, SMILES and InChIKey IDs of
all the compounds present in the first dataset. See
\code{\link{chemdiv}} for details on obtaining SMILES and InChIKey IDs.
}
\examples{
data(minimalCompData)
data(minimalSampData)
chemDivCheck(minimalCompData, minimalSampData) # Correct format
chemDivCheck(minimalCompData[c(2,3,1),], minimalSampData) # Incorrect format
}
