% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDis.R
\name{compDis}
\alias{compDis}
\title{Calculate compound dissimilarities}
\usage{
compDis(
  compoundData,
  type = c("NPClassifier", "PubChemFingerprint", "fMCS"),
  npcTable = NULL,
  unknownCompoundsMean = FALSE
)
}
\arguments{
\item{compoundData}{Data frame with the chemical compounds of interest.
Should have a column named "compound" with common names, and a column named
"smiles" with SMILES IDs for the compounds.}

\item{type}{Type of dissimilarity to calculate, NPClassifier,
PubChemFingerprint or fMCS. If more than one is choosen, a mean of
the included types will also be calculated}

\item{npcTable}{An already generated \code{\link{NPCTable}}
can optionally be supplied}

\item{unknownCompoundsMean}{If unknown compound should be given
mean dissimilarity values. If not, these will have dissimilarity 1 to
all other compounds}
}
\value{
List with compound dissimilarity matrices
}
\description{
Function to calculate dissimilarities between compounds
}
\details{
This function calculates compound dissimilarities in three different ways:
\enumerate{
\item Using the classification from NPClassifier, which largely correspond
to biosynthetic groups/pathways.
\item Using PubChem Fingerprints, a a binary substructure fingerprint with
881 binary variables describing the chemical structure of a compound.
\item fMCS, flexible Maximum Common Substructure,
a pairwise graph matching concept allowing for atom and/or
bond mismatches. Dissimilarities using NPClassifier and PubChem Fingerprints
are generated by calculating Jaccard dissimilarities from a 0/1 table with
compounds as rows and group (NPClassifier) or binary fingerprint variable
(PubChem Fingerprints) as columns. fMCS generates dissimilarity
values directly (allowing for one atom and one bond mismatch).
}
}
\examples{
data(minimalCompData)
compDis(minimalCompData)
}
