% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{funcHillDiv}
\alias{funcHillDiv}
\title{Functional Hill diversity}
\usage{
funcHillDiv(data, Dij, q)
}
\arguments{
\item{data}{Dataset. Single row in dataframe. Proportional or absolute data.}

\item{Dij}{Dissimilarity matrix.}

\item{q}{Diversity order.}
}
\value{
Functional Hill diversity (FD).
}
\description{
Function to calculate Functional Hill diversity, as in equations
4b and 6b in Chiu & Chao 2014 Plos One.
Note that this is done for one sample (row).
}
