% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalCompData}
\alias{minimalCompData}
\title{Minimal compound dataset}
\format{
A data frame with 3 rows and 3 columns. Each row is a phytochemical
compound. First column is a common name of the compound, second column is
the SMILES (Simplified Molecular-Input Line-Entry System) specification,
third column is the InChIKey (International Chemical Identifier).
}
\usage{
minimalCompData
}
\description{
A small dataset with three phytochemical compounds.
}
\keyword{datasets}
