% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalMolNet}
\alias{minimalMolNet}
\title{Minimal molecular network}
\format{
A tbl_graph with 3 nodes and 4 edges
}
\usage{
minimalMolNet
}
\description{
A minimal molecular network. Generated by the \code{\link{molNet}} function
}
\keyword{datasets}
