% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalNPCTable}
\alias{minimalNPCTable}
\title{Minimal NPClassifier table}
\format{
A dataframe with 3 compounds and their NPClassifier classifications.
}
\usage{
minimalNPCTable
}
\description{
A table with the NPClassifier pathways, superclasses and classes,
along with the compound names, smiles and inchikey. Generated by
the \code{\link{NPCTable}} function used on the
\code{\link{minimalCompData}} dataset.
}
\keyword{datasets}
