% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalSampDis}
\alias{minimalSampDis}
\title{Minimal sample dissimilarity matrix}
\format{
A 4x4 sample dissimilarity matrix
}
\usage{
minimalSampDis
}
\description{
A minimal matrix with sample dissimilarities calculated from
"minimalSampData" and minimalCompDis
using UniFracs with alpha = 0.5
}
\keyword{datasets}
