% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molNetPlot.R
\name{molNetPlot}
\alias{molNetPlot}
\title{Plot molecular network}
\usage{
molNetPlot(
  sampleData,
  networkObject,
  groupData = NULL,
  npcTable = NULL,
  plotNames = FALSE
)
}
\arguments{
\item{sampleData}{Dataframe with samples as rows and compounds as columns.}

\item{networkObject}{tidygraph network object, as created by molNet function..
Note that this is only the network object, which has to be taken
from the list.}

\item{groupData}{Grouping data. If supplied, a separate network will be
created for each group}

\item{npcTable}{An NPCTable can optionally be supplied. This will result
in network nodes being coloured by their NPC pathway classification}

\item{plotNames}{Indicates if compounds names should be included
in the molecular network plot.}
}
\value{
Molecular network(s) created with ggraph
}
\description{
Function to conveniently create a basic plot of the molecular network created
by the molNet function (based on the compound dissimilarity matrix).
}
\details{
The network object and sample dataset have to be supplied. In addition,
groupData and/or NPCTable must be supplied. If groupData is supplied,
one network will be created for each group, with node colours represtnging
the proportional concentration of the compounds. If and NPCTable is supplied,
node colours will represent NPC pathways, and node size the proportional
concentration of the compounds. Edge widths represent compound similarity,
and only edges with similarity values above the molNet function's cut-off
will be plotted
}
\examples{
data(minimalSampData)
data(minimalMolNet)
data(minimalNPCTable)
groups <- c("A", "A", "B", "B")
molNetPlot(minimalSampData, minimalMolNet)
molNetPlot(minimalSampData, minimalMolNet, plotNames = TRUE)
molNetPlot(minimalSampData, minimalMolNet, groups)
molNetPlot(minimalSampData, minimalMolNet, minimalNPCTable)
molNetPlot(minimalSampData, minimalMolNet, minimalNPCTable, plotNames = TRUE)
molNetPlot(minimalSampData, minimalMolNet, groups, minimalNPCTable)
}
