% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPCTable.R
\name{NPCTable}
\alias{NPCTable}
\title{NPClassifier Data Table}
\usage{
NPCTable(compoundData)
}
\arguments{
\item{compoundData}{Data frame with the chemical compounds of interest.
Should include a column named "compound" with common names of
the compounds and a column named "smiles" with SMILES IDs of the compounds.}
}
\value{
Data frame with the NPClassifier classification for each compound
as pathway, superclass and class. Note that compounds may be classified
in more than one group, or no group, at each level of classification.
}
\description{
This function classifies compounds with NPClassifier (Kim et al. 2021),
and puts the results in a dataframe containing the pathway,
superclass and class for each compound.
}
\details{
The function classifies natural products (phytochemical compounds)
into groups largely corresponding to biosynthetic pathways,
at three hierarchical levels: pathway, superclass
and class. This is done using the NPClassifier tool (Kim et al. 2021).
Connects to \url{https://npclassifier.ucsd.edu/}.
}
\examples{
data(minimalCompData)
NPCTable(minimalCompData)

\dontrun{
data(alpinaCompData)
NPCTable(compoundData = alpinaCompData)
}
}
\references{
Kim HW, Wang M, Leber CA, Nothias L-F, Reher R, Kang KB,
van der Hooft JJJ, Dorrestein PC, Gerwick WH, Cottrell GW. 2021.
NPClassifier: A Deep Neural Network-Based Structural Classification
Tool for Natural Products. Journal of Natural Products 84: 2795-2807.
}
