% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDis.R
\name{sampleDis}
\alias{sampleDis}
\title{Calculate sample dissimilarities}
\usage{
sampleDis(sampleData, compDisMat = NULL, type = "BrayCurtis", alpha = 1)
}
\arguments{
\item{sampleData}{Data frame with samples as rows and compounds as columns.}

\item{compDisMat}{Compound dissimilarity matrix, as calculated by
\code{\link{compDis}}. If this is supplied, Generalized UniFrac
dissimilarities can be calculated.}

\item{type}{Type of sample dissimilarity to be calculated. This is
Bray-Curtis dissimilarity, \code{BrayCurtis} and/or Generalized UniFrac
dissimilarity, \code{GenUniFrac}.}

\item{alpha}{Alpha value used to calculate Generalized UniFracs. alpha can
be set between 0 and 1. With alpha = 0 equal emphasis is put on every
branch in the dendrogram. With values closer to 1, more emphasis is put
on high abundance branches. alpha = 0.5 strikes a balance.
alpha 0.5 or 1 is recommended, with alpha = 1 as default.}
}
\value{
List with sample dissimilarity matrices. A list is always
outputted, even if only one matrix is calculated.
}
\description{
Function to calculate dissimilarities between samples.
This is either Bray-Curtis dissimilarities and/or Generalized UniFrac
dissimilarities.
}
\details{
\code{sampleDis} calculates a dissimilarity matrix for all the samples
in \code{sampleData}. Bray-Curtis dissimilarities are calculated using only
the \code{sampleData}. If a compound dissimilarity matrix, \code{compDisMat},
is supplied, Generalized UniFrac dissimilarities can also be calculated,
which uses the compound dissimilarity matrix (transformed into a dendrogram)
in the dissimilarity calculations. Thereby, the biosynthetic/structural
properties of the compounds is taken into account for the calculations
of sample dissimilarities.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
sampleDis(minimalSampData)
sampleDis(sampleData = minimalSampData, compDisMat = minimalCompDis,
type = c("BrayCurtis", "GenUniFrac"), alpha = 0.5)
}
\references{
Bray-Curtis ref?
Chen, J., K. Bittinger, E. S. Charlson, C. Hoffmann, J. Lewis, G. D. Wu,
R. G. Collman, F. D. Bushman, and H. Li. 2012.
Associating microbiome composition with environmental covariates using
generalized UniFrac distances. Bioinformatics 28:2106–2113.
}
