% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test-function}
\usage{
test_function(input)
}
\arguments{
\item{input}{Some text input}
}
\value{
Input and some more text
}
\description{
This is only a test function
}
\examples{
test_function("This input")
}
