% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpinaSampData}
\alias{alpinaSampData}
\title{Arabis alpina floral scent data}
\format{
A data frame with 87 rows and 15 columns. Each row is a sample,
each column is a floral scent compound.
}
\source{
Petren H, Torang P, Agren J, Friberg M. 2021.
Evolution of floral scent in relation to self-incompatibility and capacity
for autonomous self-pollination in the perennial herb \emph{Arabis alpina}.
Annals of Botany 127: 737-747.
}
\usage{
alpinaSampData
}
\description{
A dataset with proportional floral scent data from three populations of
the plant Arabis alpina.
}
\keyword{datasets}
