% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiv.R
\name{calcDiv}
\alias{calcDiv}
\title{Calculate various diversity/evenness measures}
\usage{
calcDiv(sampleData, compDisMat = NULL, type = "HillDiv", q = 1)
}
\arguments{
\item{sampleData}{Data frame with the relative concentration of each
compound (column) in every sample (row).}

\item{compDisMat}{Compound dissimilarity matrix, as calculated by
\code{\link{compDis}}. Has to be supplied for calculations of
Functional Hill Diversity and Rao's Q.}

\item{type}{Type(s) of diversity or evenness to calculate. Any of
\code{"HillDiv", "FuncHillDiv", "Shannon", "Simpson", "PielouEven",
"HillEven", "RaoQ"}.}

\item{q}{Diversity order to use for Hill diversity, Functional Hill
Diversity and Hill Evenness. q should be equal to or larger than zero.
This parameter determines the sensitivity of the (Functional) Hill Diversity
measure to the relative frequencies of compounds. Commonly set to 0, 1 or 2,
although any value > 0 may be used. For q = 0 compound proportions are not
taken into account. For q = 1 (default) compounds are weighed according to
their proportion in the sample. For q = 2, more weight is put on compounds
with high proportions.}
}
\value{
Data frame with calculated diversity/evenness values for each sample.
}
\description{
Function to calculate different common measures of diversity and evenness.
This includes Hill diversity, Functional Hill diversity,
Shannon's diversity, Simpson diversity, Rao's Q, Pielou's evenness and
Hill evenness.
}
\details{
The function can calculate the following values:
\itemize{
\item \code{Shannon}. Shannon's Diversity.
\item \code{Simpson}. Simpson Diversity, often referred to as
the Inverse Simpson Index.
\item \code{HillDiv}. Hill Diversity. Also referred to as the Hill number
or the effective number of species (here compounds). The parameter q
determines the sensitivity of the measure to the relative frequencies of
compounds (see above for details). For q = 0, this equals the number of
compounds in a sample. For q = 1, this equals the exponential of
Shannon's Diversity. For q = 2, this equals the (inverse) Simpson Diversity.
\item \code{FuncHillDiv}. Functional Hill Diversity. Requires a compound
dissimilarity matrix. Functional Hill Diversity quantifies the effective
total dissimilarity between compounds in the sample. The parameter q
determines the  sensitivity of the measure to the relative frequencies
of compounds (see above for details). For q = 0, this is
equal to Functional Attribute Diversity (FAD) which is the sum of all
dissimilarities in the dissimilarity matrix.
\item \code{RaoQ}. Rao's Q. The perhaps most common measure of functional
diversity. Requires a compound dissimilarity matrix. This represents the
average dissimilarity of two randomly selected (weighed by their proportions)
compounds in the sample.
\item \code{PielouEven}. Pielou's Evenness. The perhaps most common
evenness measure. Equal to the Shannon's Diversity divided by the
natural logarithm of the number of compounds. This is a relative measure
that is not replication invariant.
\item \code{HillEven}. Hill Evenness. This is equal to the Hill Diversity
(at the given value of q) divided by the number of compounds.
This is a replication invariant measure.
}
}
\examples{
data(minimalSampData)
data(minimalCompDis)
calcDiv(sampleData = minimalSampData)
calcDiv(sampleData = minimalSampData, type = c("HillDiv", "HillEven"))
calcDiv(sampleData = minimalSampData, compDisMat = minimalCompDis,
type = "FuncHillDiv", q = 2)
}
\references{
Chao's and other papers, e.g. Botta-Dukat 2005, Hill 1973,
Pielous paper, Tuomisto oikos, Chiu & Chao, Shannon/Simpson papers?
Chao, A., C.-H. Chiu, and L. Jost. 2014. Unifying Species Diversity,
Phylogenetic Diversity, Functional Diversity, and Related Similarity
and Differentiation Measures Through Hill Numbers. Annual Review of
Ecology, Evolution, and Systematics 45:297–324.
}
