% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDiv.R
\name{calcDiv}
\alias{calcDiv}
\title{Calculate various diversity/evenness measures}
\usage{
calcDiv(sampleData, compDisMat = NULL, type = "HillDiv", q = 1)
}
\arguments{
\item{sampleData}{Dataframe with samples as rows and compounds as columns.}

\item{compDisMat}{Compound distance matrix. Has to be supplied for
calculations of Functional Hill diversity or Rao's Q.}

\item{type}{Type of diversity or evenness to calculate. Any of
\code{"HillDiv", "FuncHillDiv", "Shannon", "Simpson", "PielouEven",
"HillEven", "RaoQ"}.}

\item{q}{Diversity order to use for (Functional) Hill diversity}
}
\value{
Data frame with calculated diversity/evenness for each sample
}
\description{
Function to calculate Hill diveristy, Functional Hill diversity,
Shannon's diveristy, Simpson diveristy, Pielou's evenness, Hill evenness
and Rao's Q.
}
\details{
Add details here
Hill diversity:
Functional Hill diversity:
Shannon's diversity:
Simpson diversity:
Pielou's evenness:
Hill evenness:
Rao's Q:
}
\examples{
\dontrun{Add minimal executable example when PubChem PUG works}
}
