% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemDivPlot.R
\name{chemDivPlot}
\alias{chemDivPlot}
\title{Plot chemodiversity}
\usage{
chemDivPlot(
  compDisMat = NULL,
  divData = NULL,
  divProfData = NULL,
  sampleDisMat = NULL,
  groupData = NULL
)
}
\arguments{
\item{compDisMat}{Compound dissimilarity matrix,
generated by the \code{compDis} function.}

\item{divData}{Diversity/evenness data frame,
generated by the \code{calcDiv} function. This dataframe can contain
a single or multiple columns with diversity/evenness measures.}

\item{divProfData}{Diversity profile data frame,
generated by the \code{calcDivProf} function.}

\item{sampleDisMat}{Sample dissimilarity matrix, generated by
the \code{sampleDis} function. This can be either the list of one or
both matrices outputted by the function, or a single matrix directly.}

\item{groupData}{Grouping data.}
}
\value{
Various plots on chemodiversity.
}
\description{
Function to conveniently create basic plots of the different types of
chemodiversity measurements created by various functions in the package.
}
\details{
The function can create four different types of plots (using ggplot2)
depending on which input data is supplied.
\itemize{
\item A compound dissimilarity matrix, \code{compDisMat}, will be
plotted as a dendrogram visualizing which compounds are
structurally/biosynthethically similar to each other.
\item Diversity/evenness, \code{divData}, is plotted as a boxplot.
\item A diversity profile, plotting (Functional) Hill diversity at different
values of q will be plotted with \code{divProfData} (NB! This is
the whole list outputted by calcDivProf, as qMin and qMax are used too).
\item A sample dissimilarity matrix, \code{sampleDisMat}, will be plotted
as an NMDS plot.
\item Grouping data, \code{groupData}, may be supplied.
}
Note that this function can take any combination of the four arguments
as input, and argument names should always be specified to ensure
each dataset is correctly plotted.
}
\examples{
data(minimalCompDis)
data(minimalSampDis)
minimalDiv <- calcDiv(minimalSampData, minimalCompDis, type = "FuncHillDiv")
minimalDivProf <- calcDivProf(minimalSampData, minimalCompDis, type = "FuncHillDiv")
groups <- c("A", "A", "B", "B")
chemDivPlot(compDisMat = minimalCompDis, divData = minimalDiv,
divProfData = minimalDivProf, sampleDisMat = minimalSampDis,
groupData = groups)
}
