% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compDis.R
\name{compDis}
\alias{compDis}
\title{Calculate compound dissimilarities}
\usage{
compDis(
  compoundData,
  type = c("NPClassifier", "PubChemFingerprint", "fMCS"),
  npcTable = NULL,
  unknownCompoundsMean = FALSE
)
}
\arguments{
\item{compoundData}{Data frame with the chemical compounds of interest.
Should have a column named "compound" with common names, and a column named
"smiles" with SMILES IDs for the compounds.}

\item{type}{Type of dissimilarity to calculate, NPClassifier,
PubChemFingerprint or fMCS. If more than one is choosen, a mean of
the included types will also be calculated.}

\item{npcTable}{An already generated \code{\link{NPCTable}}
can optionally be supplied. If compound distances are to be calculated using
the NPC-classification and NPClassifier is unable to fully classify
all compounds, it might be beneficial to run \code{\link{NPCTable}},
manually edit the resulting data frame with probable classifications if
possible (from the SI in Kim et al. 2021), and then supply this
classification to \code{compDis()} with the npcTable argument. This will ensure
that compound dissimilarities are computed optimally.}

\item{unknownCompoundsMean}{If unknown compounds, i.e. ones without SMILES,
should be given mean dissimilarity values. If not, these will have
dissimilarity 1 to all other compounds.}
}
\value{
List with compound dissimilarity matrices. A list is always
outputted, even if only one matrix is calculated. Downstream functions,
including \code{\link{calcDiv}}, \code{\link{calcBetaDiv}},
\code{\link{calcDivProf}}, \code{\link{sampleDis}},
\code{\link{molNet}} and \code{\link{chemDivPlot}} require only the matrix
as input (subsetted with e.g. fullList$specificMatrix) and not the whole list
}
\description{
Function to calculate dissimilarities between compounds.
}
\details{
This function calculates compound dissimilarities in three different ways:
\enumerate{
\item Using the classification from NPClassifier, which largely correspond
to biosynthetic groups/pathways. Classifications are fetched from
\url{https://npclassifier.ucsd.edu/}.
\item Using PubChem Fingerprints, a binary substructure fingerprint with
881 binary variables describing the chemical structure of a compound.
See \url{https://pubchemdocs.ncbi.nlm.nih.gov/data-specification} for
information on PubChem Fingerprints.
\item fMCS, flexible Maximum Common Substructure,
a pairwise graph matching concept allowing for atom and/or
bond mismatches (Wang et al 2013). While potentially a very accurate
similarity measure, fMCS is more computationally demanding
than the other methods, and will take a significant amount of time
for larger data sets.
Dissimilarities using NPClassifier and PubChem Fingerprints
are generated by calculating Jaccard dissimilarities from a 0/1 table with
compounds as rows and group (NPClassifier) or binary fingerprint variable
(PubChem Fingerprints) as columns. fMCS generates dissimilarity
values directly (allowing for one atom and one bond mismatch).
}
}
\examples{
data(minimalCompData)
compDis(minimalCompData)
}
\references{
Kim, H. W., M. Wang, C. A. Leber, L.-F. Nothias, R. Reher, K. B. Kang,
J. J. J. van der Hooft, P. C. Dorrestein, W. H. Gerwick,
and G. W. Cottrell. 2021. NPClassifier: A Deep Neural Network-Based
Structural Classification Tool for Natural Products.
Journal of Natural Products:acs.jnatprod.1c00399.
UPDATE WITH FULL REF WHEN AVAILABLE

Wang, Y., T. W. H. Backman, K. Horan, and T. Girke. 2013.
fmcsR: mismatch tolerant maximum common substructure searching in R.
Bioinformatics 29:2792-2794.
}
