% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalCompDis}
\alias{minimalCompDis}
\title{Minimal compound dissimilarity matrix}
\format{
A 3x3 compound dissimilarity matrix.
}
\usage{
minimalCompDis
}
\description{
A matrix with compound dissimilarities calculated using
\code{\link{compDis}} with \code{PubChemFingerprint} as \code{type},
for the compounds in \code{\link{minimalCompData}}.
}
\keyword{datasets}
