% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalMolNet}
\alias{minimalMolNet}
\title{Minimal molecular network}
\format{
A \code{tbl_graph} object with 3 nodes and 4 edges.
}
\usage{
minimalMolNet
}
\description{
A molecular network. Generated by the \code{\link{molNet}} function
used on the \code{\link{minimalCompDis}} dataset, with
\code{cutOff = "median"}.
}
\keyword{datasets}
