% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalSampData}
\alias{minimalSampData}
\title{Minimal sample dataset}
\format{
A data frame with 4 rows and 4 columns. Each row is a sample,
first column is group, the other columns are scent compounds.
}
\usage{
minimalSampData
}
\description{
A minimal made up dataset with floral scent data.
}
\keyword{datasets}
