% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalSampData}
\alias{minimalSampData}
\title{Minimal sample dataset}
\format{
A data frame with 4 rows and 3 columns. Each row is a sample,
each column is a phytochemical compound.
}
\usage{
minimalSampData
}
\description{
A small made up dataset with phytochemical data.
}
\keyword{datasets}
