% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minimalSampDis}
\alias{minimalSampDis}
\title{Minimal sample dissimilarity matrix}
\format{
A 4x4 sample dissimilarity matrix.
}
\usage{
minimalSampDis
}
\description{
A matrix with sample dissimilarities calculated from
\code{\link{minimalSampData}} and \code{\link{minimalCompDis}}
using \code{\link{sampleDis}} with Generalized UniFracs with
\code{alpha = 0.5}.
}
\keyword{datasets}
