% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molNet.R
\name{molNet}
\alias{molNet}
\title{Generate molecular network with properties}
\usage{
molNet(compDisMat, npcTable = NULL, cutOff = "median")
}
\arguments{
\item{compDisMat}{Compound dissimilarity matrix, as calculated by
\code{\link{compDis}}.}

\item{npcTable}{An already generated \code{\link{NPCTable}} can be supplied
for calculations of the number of NPC pathways and network modularity.}

\item{cutOff}{Cut-off value for compound similarities. Any similarity
lower than this value will be set to zero when the network is generated,
which strongly affects the look of the network. The value can be set
manually 0 <= cutOff <= 1 (maybe mention that others use 0.6 for another
kind of molecular similarity measure), to the \code{median} dissimilarity
value in the compDisMat, or to \code{minPathway}, the lowest within-pathway
similarity (which allows all within-NPC-pathway to be kept) if an NPCTable
is supplied.}
}
\value{
List with a graph (tidygrph/igraph?) object, the number of compounds,
number of NPC pathways and modularity.
}
\description{
Function which generates a molecular network object, and some
basic properties of the network.
}
\details{
Note that the supplied dissimilarity matrix is translated to a
similarity matrix, and this is what cutOff values are set for.
}
\examples{
data(minimalCompDis)
data(minimalNPCTable)
molNet(minimalCompDis)
molNet(minimalCompDis, minimalNPCTable, cutOff = 0)

\dontrun{
data(alpinaCompData)
alpinaNPCTable <- NPCTable(compoundData = alpinaCompData)
alpinaCompDis <- compDis(compoundData = alpinaCompData)
molNet(compDisMat = alpinaCompDis$fingerDisMat,
npcTable = alpinaNPCTable, cutOff = 0.75)
}
}
