% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molNetPlot.R
\name{molNetPlot}
\alias{molNetPlot}
\title{Plot molecular network}
\usage{
molNetPlot(
  sampleData,
  networkObject,
  groupData = NULL,
  npcTable = NULL,
  plotNames = FALSE,
  layout = "kk"
)
}
\arguments{
\item{sampleData}{Data frame with the relative concentration of each
compound (column) in every sample (row).}

\item{networkObject}{A network object, as created by the
\code{\link{molNet}} function. Note that this is only the network object,
which is one of the elements in the list outputted by \code{\link{molNet}}.
The network is extracted as molNetOutput$networkObject.}

\item{groupData}{Grouping data. If supplied, a separate network will be
created for each group. Should be either a vector, or a data frame
with a single column.}

\item{npcTable}{It is recommended but optional to supply an
\code{\link{NPCTable}} This will result in network nodes being coloured
by their NPC pathway classification.}

\item{plotNames}{Indicates if compounds names should be included
in the molecular network plot.}

\item{layout}{Layout used by \code{\link[ggraph]{ggraph}} when creating
the network. The default chosen here, \code{"kk"}, is the the Kamada-Kawai
layout algorithm which in most cases should produce a visually
pleasing network. Another useful option is \code{"circle"}, which puts all
nodes in a circle, for easier comparisons between different networks.}
}
\value{
Molecular network(s) created using \code{ggraph}.
}
\description{
Function to conveniently create a basic plot of the molecular network
created by the \code{\link{molNet}} function. Molecular networks can be
used to illustrate the biosynthetic/structural similarity of
phytochemical compounds in a sample, while simultaneously visualizing
their relative concentrations. Nodes are compounds,
with node size or node colour representing proportional concentration
of the compounds. Edge widths represent compound similarity.
}
\details{
The network object from \code{\link{molNet}} and sample data frame have to
be supplied. In addition, groupData and/or an \code{\link{NPCTable}} can
be supplied. If groupData is supplied, one network will be created
for each group. If an NPCTable is supplied, which is recommended,
node colours will represent NPC pathways, and node size the proportional
concentration of the compounds. Edge widths represent compound similarity,
and only edges with similarity values above the \code{\link{molNet}}
function's cut-off will be plotted.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
data(minimalNPCTable)
data(minimalMolNet)
groups <- c("A", "A", "B", "B")
molNetPlot(minimalSampData, minimalMolNet)
molNetPlot(minimalSampData, minimalMolNet, groups)
molNetPlot(minimalSampData, minimalMolNet, npcTable = minimalNPCTable)
molNetPlot(minimalSampData, minimalMolNet, plotNames = TRUE)

\dontrun{
data(alpinaCompData)
data(alpinaSampData)
data(alpinaPopData)
alpinaNPCTable <- NPCTable(compoundData = alpinaCompData)
alpinaCompDis <- compDis(compoundData = alpinaCompData)
alpinaMolNet <- molNet(compDisMat = alpinaCompDis$fingerDisMat,
npcTable = alpinaNPCTable, cutOff = 0.75)
molNetPlot(sampleData = alpinaSampData,
networkObject = alpinaMolNet$networkObject,
groupData = alpinaPopData, npcTable = alpinaNPCTable)
}
}
