% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molNetPlot.R
\name{molNetPlot}
\alias{molNetPlot}
\title{Plot molecular network}
\usage{
molNetPlot(
  sampleData,
  networkObject,
  groupData = NULL,
  npcTable = NULL,
  plotNames = FALSE
)
}
\arguments{
\item{sampleData}{Data frame with samples as rows and compounds as columns.}

\item{networkObject}{tidygraph network object, as created by the
\code{\link{molNet}} function. Note that this is only the network object,
which is one of the elements in the list outputted by \code{\link{molNet}}.
The network is extracted as molNetOutput$networkObject.}

\item{groupData}{Grouping data. If supplied, a separate network will be
created for each group.}

\item{npcTable}{An \code{\link{NPCTable}} can optionally be supplied.
This will result in network nodes being coloured by their
NPC pathway classification.}

\item{plotNames}{Indicates if compounds names should be included
in the molecular network plot.}
}
\value{
Molecular network(s) created with ggraph.
}
\description{
Function to conveniently create a basic plot of the molecular network
created by the \code{\link{molNet}} function
(based on the compound dissimilarity matrix). Nodes represent compounds,
with size proportional to proportions, and edge widths represent
compound similarity.
}
\details{
The network object from \code{\link{molNet}}  and sample dataset have to
be supplied. In addition, groupData and/or \code{\link{NPCTable}} can
be supplied. If groupData is supplied,
one network will be created for each group. If and NPCTable is supplied,
node colours will represent NPC pathways, and node size the proportional
concentration of the compounds. Edge widths represent compound similarity,
and only edges with similarity values above the \code{\link{molNet}}
function's cut-off will be plotted.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
data(minimalNPCTable)
groups <- c("A", "A", "B", "B")
minimalMolNet <- molNet(minimalCompDis)
molNetPlot(minimalSampData, minimalMolNet$networkObject)
molNetPlot(minimalSampData, minimalMolNet$networkObject, groups)
molNetPlot(minimalSampData,
minimalMolNet$networkObject,
npcTable = minimalNPCTable)
molNetPlot(minimalSampData,
minimalMolNet$networkObject,
plotNames = TRUE)
}
