% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPCTable.R
\name{NPCTable}
\alias{NPCTable}
\title{NPClassifier Data Table}
\usage{
NPCTable(compoundData)
}
\arguments{
\item{compoundData}{Data frame with the chemical compounds of interest.
Should have a column named "compound" with common names, and a column named
"smiles" with isomeric or canonical SMILES IDs for the compounds.}
}
\value{
Data frame with the NPClassifier classification for each compound
as pathways, superclass and class. Note that compounds may be classified
in more than one group, or no group, at each level of classification.
}
\description{
This function classifies compounds with NPClassifier,
and puts the results in a dataframe containing the pathway,
superclass and class for each compound.
}
\details{
NPCTable classifies compounds into groups largely corresponding to
biosynthesis pathways (at three hierarchical levels: pathway, superclass,
class) using thr NPClassifier tool (Kim et al. 2021).
Connects to \url{https://npclassifier.ucsd.edu/}.
}
\examples{
data(minimalCompData)
NPCTable(minimalCompData)
}
\references{
Kim, H. W., M. Wang, C. A. Leber, L.-F. Nothias, R. Reher, K. B. Kang,
J. J. J. van der Hooft, P. C. Dorrestein, W. H. Gerwick,
and G. W. Cottrell. 2021. NPClassifier: A Deep Neural Network-Based
Structural Classification Tool for Natural Products.
Journal of Natural Products:acs.jnatprod.1c00399.
UPDATE WITH FULL REF WHEN AVAILABLE
}
