% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickChemDiv.R
\name{quickChemDiv}
\alias{quickChemDiv}
\title{Quickly calculate/plot chemodiversity}
\usage{
quickChemDiv(compoundData, sampleData, groupData = NULL, outputType = "plots")
}
\arguments{
\item{compoundData}{Data frame with the chemical compounds of interest.
Should have a column named "compound" with common names, a column named
"smiles" with SMILES IDs for the compounds, and a column named "inchikey"
with the InChIKey IDs for the compounds. See \code{\link{chemdiv}} for
details on obtaining SMILES and InChIKey IDs.}

\item{sampleData}{Data frame with the relative concentration of each
compound (column) in every sample (row).}

\item{groupData}{Grouping data.}

\item{outputType}{Type of output that should be returned: either
\code{"data"} to output a list with different types of chemodiversity data,
or \code{"plots"} to instead produce standard plots of this data.}
}
\value{
Four types of chemodiversity measures, as either elements in a list
or separate plots. If \code{outputType = "data"}, function returns
a compound dissimilarity matrix, a data frame with Functional Hill
Diversity at q = 1, a data frame with a Functional Hill Diversity profile
for q = 0-3, and a sample dissimilarity matrix. If
\code{outputType = "plots"}, these data sets are plotted as a dendrogram,
a boxplot, a diversity profile plot and an NMDS plot, respectively.
}
\description{
This function is a shortcut that makes use of many of the other functions in
the package. In one simple step chemodiversity is calculated, and if
requested also visualized, for users wanting to quickly explore their
data using standard parameters.
}
\details{
The function takes compound data and sample data as input. See
\code{\link{chemdiv}} for details on data formatting. This function then
uses the following other functions in the package:
\itemize{
\item \code{\link{compDis}} is used to calculate compound dissimilarities
using PubChem Fingerprints.
\item \code{\link{calcDiv}} is used to calculate Functional Hill Diversity
for q = 1.
\item \code{\link{calcDivProf}} is used to calculate a diversity profile
with Functional Hill Diversity for q = 0-3.
\item \code{\link{sampleDis}} is used to calculate Generalized UniFrac
dissimilarities between samples.
\item \code{\link{chemDivPlot}} is used to create four different
chemodiversity plots if requested.
}

\code{quickChemDiv} is designed to provide an easy way to visualize
the most central measures of phytochemical diversity. It uses
default parameters to do so, which should be reasonable in most cases.
However, for detailed analyses we recommend using the separate
functions to allow for full control of function parameters and output.
}
\examples{
data(minimalCompData)
data(minimalSampData)
groups <- c("A", "A", "B", "B")
quickChemDiv(compoundData = minimalCompData, sampleData = minimalSampData,
groupData = groups, outputType = "plots")
}
