% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDis.R
\name{sampleDis}
\alias{sampleDis}
\title{Calculate sample dissimilarities}
\usage{
sampleDis(sampleData, compDisMat = NULL, alpha = 1)
}
\arguments{
\item{sampleData}{Dataframe with samples as rows and compounds as columns.}

\item{compDisMat}{Compound distance matrix. If this is supplied UniFrac
dissimilarities are calcualted, otherwise Bray-Curtis distances
are calculated}

\item{alpha}{Alpha value used to calculate UniFracs}
}
\value{
Sample dissimilarity matrix with Bray-Curtis or
UniFrac dissimilarities
}
\description{
Function to calculate dissimilarities between samples.
If only sample data is supplied Bray-Curtis dissimilarities are calculated.
If also a compound dissimilarity matrix is supplied, Generalized UniFrac
dissimilarities are calculated.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
sampleDis(minimalSampData)
sampleDis(minimalSampData, minimalCompDis, alpha = 0.5)
}
