% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDis.R
\name{sampleDis}
\alias{sampleDis}
\title{Calculate sample dissimilarities}
\usage{
sampleDis(sampleData, compDisMat = NULL, alpha = 1)
}
\arguments{
\item{sampleData}{Data frame with samples as rows and compounds as columns.}

\item{compDisMat}{Compound dissimilarity matrix, as calculated by
\code{\link{compDis}}. If this is supplied UniFrac dissimilarities are
calculated, otherwise Bray-Curtis distances are calculated}

\item{alpha}{Alpha value used to calculate UniFracs. alpha can be set
between 0 and 1. With alpha = 0 equal emphasis is put on every branch,
With values closer to 1, more emphasis is put on high abundance branches.
alpha = 0.5 strikes a balance. alpha 0.5 or 1 is recommended}
}
\value{
Sample dissimilarity matrix with Bray-Curtis or
UniFrac dissimilarities
}
\description{
Function to calculate dissimilarities between samples.
If only sample data is supplied Bray-Curtis dissimilarities are calculated.
If also a compound dissimilarity matrix is supplied, Generalized UniFrac
dissimilarities are calculated.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
sampleDis(minimalSampData)
sampleDis(minimalSampData, minimalCompDis, alpha = 0.5)
}
