% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleDis.R
\name{sampleDis}
\alias{sampleDis}
\title{Calculate sample dissimilarities}
\usage{
sampleDis(sampleData, compDisMat = NULL, type = "BrayCurtis", alpha = 1)
}
\arguments{
\item{sampleData}{Data frame with samples as rows and compounds as columns.}

\item{compDisMat}{Compound dissimilarity matrix, as calculated by
\code{\link{compDis}}. If this is supplied, Generalized UniFrac
dissimilarities can be calculated.}

\item{type}{Type of sample dissimilarity to be calculated. This is
Bray-Curtis dissimilarity, \code{BrayCurtis} and/or Generalized UniFrac
dissimilarity, \code{GenUniFrac}.}

\item{alpha}{Parameter used in calculations of Generalized UniFracs
dissimilarities. alpha can be set between 0 and 1.
With alpha = 0, equal weight is put on every
branch in the dendrogram. With alpha = 1, branches are weighted by their
abundance, and hence more emphasis is put on high abundance branches.
alpha = 0.5 strikes a balance between the two.
alpha 0.5 or 1 is recommended, with alpha = 1 as default.}
}
\value{
List with sample dissimilarity matrices. A list is always
outputted, even if only one matrix is calculated.
}
\description{
Function to calculate dissimilarities between samples.
This is either Bray-Curtis dissimilarities and/or Generalized UniFrac
dissimilarities.
}
\details{
\code{sampleDis} calculates a dissimilarity matrix for all the samples
in \code{sampleData}, for the given dissimilarity index/indices.
Bray-Curtis dissimilarities are calculated using only
the \code{sampleData}. This is the most commonly calculated dissimilarity
index used for phytochemical data (other types of dissimilarities are
easily calculated using the \code{vegdist} function in the \code{vegan}
package).

If a compound dissimilarity matrix, \code{compDisMat},
is supplied, Generalized UniFrac dissimilarities can be calculated, which
also uses the compound dissimilarity matrix for the sample dissimilarity
calculations. For the calculation of Generalized UniFrac dissimilarities,
the compound dissimilarity matrix is transformed into a dendrogram using
hierarchical clustering. Calculation of UniFrac distances
quantifies the fraction of the total branch length of the dendrogram that
leads to compounds present in either sample, but not both.
The (weighted) Generalized UniFrac dissimilarities
implemented here additionally takes compound abundances
into account. In this way, both the relative proportions of compounds and
the biosynthetic/structural properties of these compounds are accounted for
in the calculations of sample dissimilarities. As for Bray-Curtis
dissimilarities, Generalized UniFrac dissimilarities range in value from
0 to 1.
}
\examples{
data(minimalSampData)
data(minimalCompDis)
sampleDis(minimalSampData)
sampleDis(sampleData = minimalSampData, compDisMat = minimalCompDis,
type = c("BrayCurtis", "GenUniFrac"), alpha = 0.5)
}
\references{
Bray, J. R., and J. T. Curtis. 1957. An Ordination of the Upland Forest
Communities of Southern Wisconsin. Ecological Monographs 27:325-349.

Chen, J., K. Bittinger, E. S. Charlson, C. Hoffmann, J. Lewis, G. D. Wu,
R. G. Collman, F. D. Bushman, and H. Li. 2012. Associating microbiome
composition with environmental covariates using generalized UniFrac
distances. Bioinformatics 28:2106-2113.

Lozupone, C., and R. Knight. 2005. UniFrac: a New Phylogenetic Method
for Comparing Microbial Communities. Applied and Environmental
Microbiology 71:8228-8235.
}
